refactor(scores): extract function and clean code of ScoresPage Filters
