+++ plugins/leemons-plugin-academic-calendar/frontend/src/hooks/useAcademicCalendarConfig.js
@@ -1,11 +1,12 @@
 import { useQueries } from '@tanstack/react-query';
 import { getConfig } from '@academic-calendar/request/config';
 
-function useAcademicCalendarConfig(program) {
+export function useAcademicCalendarConfig(program, options) {
   const programs = Array.isArray(program) ? program : [program];
 
   const queries = useQueries({
     queries: programs.map((p) => ({
+      ...options,
       queryKey: ['useAcademicCalendarConfig', { program: p }],
       queryFn: async () => {
         const { config } = await getConfig(p);
@@ -21,4 +22,4 @@ function useAcademicCalendarConfig(program) {
   return queries[0];
 }
 
-export { useAcademicCalendarConfig };
+export default useAcademicCalendarConfig;
