+++ plugins/leemons-plugin-scores/frontend/src/components/ScoresPage/Filters/hooks/useSelectedClass.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import { useWatch } from 'react-hook-form';
+
+export default function useSelectedClass({ classes, control, classID }) {
+  const selectedClass = useWatch({
+    control,
+    name: 'class',
+    defaultValue: null,
+  });
+
+  const selectedClassId = classID || selectedClass?.[0];
+
+  return React.useMemo(() => {
+    if (!selectedClassId || !classes?.length) {
+      return null;
+    }
+
+    return classes.find((klass) => klass.id === selectedClassId);
+  }, [selectedClassId, classes]);
+}
